chore: deprecate some integration internals
