+++ ddtrace/contrib/aiobotocore/__init__.py
@@ -32,7 +32,11 @@ required_modules = ["aiobotocore.client"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.aiobotocore.patch import get_version
+        from ..internal.aiobotocore.patch import patch
 
         __all__ = ["patch", "get_version"]
