+++ ddtrace/contrib/aiohttp/__init__.py
@@ -91,9 +91,13 @@ required_modules = ["aiohttp"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .middlewares import trace_app
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
+        from ..internal.aiohttp.middlewares import trace_app
+        from ..internal.aiohttp.patch import get_version
+
+        # Expose public methods
+        from ..internal.aiohttp.patch import patch
+        from ..internal.aiohttp.patch import unpatch
 
         __all__ = ["patch", "unpatch", "trace_app", "get_version"]
