+++ ddtrace/contrib/asyncio/compat.py
@@ -1,27 +1,15 @@
-import asyncio
-
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
 from ddtrace.vendor.debtcollector import deprecate
 
+from ..internal.asyncio.compat import *  # noqa: F401,F403
 
-if hasattr(asyncio, "current_task"):
 
-    def asyncio_current_task():
-        deprecate(
-            "ddtrace.contrib.asyncio.create_task(..) is deprecated. The ddtrace library fully supports propagating"
-            " trace contextes to async tasks. No additional configurations are required.",
-            version="3.0.0",
-        )
-        try:
-            return asyncio.current_task()
-        except RuntimeError:
-            return None
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-else:
-
-    def asyncio_current_task():
-        deprecate(
-            "ddtrace.contrib.asyncio.create_task(..) is deprecated. The ddtrace library fully supports propagating"
-            " trace contextes to async tasks. No additional configurations are required.",
-            version="3.0.0",
-        )
-        return asyncio.Task.current_task()
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
