+++ tests/.suitespec.json
@@ -129,6 +129,7 @@
             "ddtrace/contrib/__init__.py",
             "ddtrace/contrib/trace_utils.py",
             "ddtrace/contrib/trace_utils_async.py",
+            "ddtrace/contrib/internal/*",
             "ddtrace/ext/__init__.py",
             "ddtrace/ext/http.py",
             "ddtrace/ext/net.py",
